WIP: Remove tls_options.xmit_hold
Also remove reliable.hold, reliable_schedule_now.
The hold members were added to support "--port-share"
(commit 6add6b2), but they are bugfixes on the TLS state
machine on the server, which starts up on TCP connect,
when it should start on a HARD_RESET message from the
client.

The reliable_schedule_now has existed since the initial
git commit 6fbf66, but it is not called anywhere.

WIP because --port-share is likely broken now. The other
part of this commit is to properly fix the TLS state machine.