Implement source tree structure versioning
In order for the test suite (which is maintained in a tree separate from
the source code) to be backwards compatible, it must be able
to find the source of the tested units even as the source tree gets
reorganized from time to time.

This commit sets the $openvpn_srcdir, $compat_srcdir appropriately,
depending on the source tree the testsuite gets merged into. When merged
into a 2.0 source tree, it points to sources in the root directory of the
project, in a 2.3 source tree it points to sources in /src/... and so on.
Makefile.am files in various unit test directories need only
"include $(top_srcdir)/tests/srcpaths.inc" to have these _srcdir variables
set correctly.