Add base64 unit test, and add sample test
Tests NDEBUG, ENABLE_DEBUG configurations, memory leaks/
unititialized memory access, and buffer over/under flow.

This commit adds the base64 "feature" to the project so that it is
integrated with the unit tests. As this feature pre-dates the
testsuite, it was not properly defined to the testsuite in
configure.ac prior to this commit. The initial test was written for
the base64 implementation from openvpn 2.0 (start of the repo).
It will be updated to support up to the current implementation in a
subsequent commit.

This commit serves as an example on how to add existing features to the
testsuite so they can be tested backwards-compatibly.

Should the base64 "feature" be removed in the future, this commit makes it
possible to retain the most up-to-date base64 tests to be used against
older versions of the project, prior to its removal. If future bugs will
be discovered, regression tests can be added to the testsuite, and they can
be run against old project versions, such that bugfixes for those versions
can be proven to be correct (ie, do not fail the up-to-date regression
tests)

Note: Since this is a legacy feature, the IMPLEMENT_base64 is defined by
detection of base64.c in the source tree. For new features, their
IMPLEMENT_* conditionals are defined initially as [false] in the test
repo, then toggled [true] in the source repo, as described in
tests/unit/README.md