Update base64 unit test for historical revisions
The first version the base64 unit test was written as a sample of the unit
test framework. At the time, it was implemented for the 2.0.3 source.
In the meanwhile, the base64 code has undergone revisions, which changed
the function signatures.

This commit demonstrates how a legacy test should be updated to detect
several implementations of the same feature, and test them all with the
same unit test.

The unit test is able to distinguish the implementations under test, and
thus, can even do additional tests based on the defines IMPLEMENTATION_*
although in this instance there are no tests that apply uniquely to one
version and not the other, that is to say, all implementations are
functionally equivalent, although their interface has changed.

For new features, the IMPLEMENTED_* conditionals in configure.ac should be
defined directly, without the need to detect them. The remaining changes
are equally relevant to maintaining one unit test compatibility with
a changing implementation.