Port tls_crypt to tests/unit framework
This is a straight-up port, maintaining the incorrect behaviour from unit_tests:
If configured with --disable-crypto, the tls_crypt unit will not be tested
Likewise, if configured with openssl, the mbedtls version is not tested,
and vice versa.

This behaviour may be appropriate for package builders, but not for maintainers.

Maintainers should test all variants. The next commit to tls_crypt updates it
to capitalize the effort made to write the test, by testing all variants.

The two commits are not squashed, in order to illustrate how a unit-test can be
written to exercise the unit thoroughly, regardless of build options.