Unit test tls_crypt against both openssl and mbedtls
When configured with --enable-unit-tests=all, both openssl and mbedtls are
tested, regardless which is used to build the application. This could be
called "maintainer mode", where every available tests are run. The goal is,
if a feature is implemented, and a test exists, it should be exercised for
every possible build configuration.

Note. This specific test is failing the ndebug test by crashing with
segfault. I believe this is due to an ASSERT that has side-effects in one
of either the UUT (tls_crypt.c), or the supporting sources.

Also added a valgrind uninitialized memory checking and memory leaks
checking, in addition to the cmocka memory leak checking.

The cmocka leak checking instrumentation is also causing some tests
to fail.

Two (temporary) tests without memory instrumentation are also present
to demonstrate that porting to the tests/unit framework was successful
even though most test modes fail at the moment. The passing tests are
protocol.nomemchecks.tls_crypt.openssl.test and
protocol.nomemchecks.tls_crypt.mbedtls.test