Add rfc4821-related buffer unit tests
rfc4821 uses some buffer features that are not used elsewhere:
buf_realloc(), buf_fill_incompressible() and the ip_pmtudisc field in the
buffer struct.

This test specifically targets these features of the buffer code.
It is built conditionally on rfc4821 also being implemented in the branch,
otherwise it's not built.

Adds only memleak and ndebug tests for regression, but no sanity test.