Implement quirk managment
Protocol Quirks/bugs do happen, with the effect that subsequent versions
of the software needs to work around bugs existing in previous versions
for the sake of interoperability.

This commit introduces the WORKAROUND(bug) macro and the infrastructure
to support it. The WORKAROUND macro is a conditional to be used in code
wherever tricky/convoluted/ugly code is written to cope with a bug
implemented in peers running older versions of the same codebase.

The WORKAROUND macro not only acts as the condition to activate workaround
functionality, but also as documentation for future developers, alerting
them to the reason a particular section of code looks unorthodox.

Another use is to enable planning and testing obsolescence of features,
or ending support for interoperation with old versions of the software.

The infrastructure is thoroughly documented in the file quirks.h

There are a number of workarounds throughout the code base that exist for
legacy purposes, but they are undocumented, and thus hard to remove, and
fragile, meaning later code changes easily break old workarounds. One
example is the "OpenVPN Configuration Control" (occ.{c,h}):

OCC exists to deal with:

1. Unimplemented SSL_shutdown handshake, (workaround is OCC_EXIT)
2. Unimplemented PMTU/buffering management (workaround is OCC_*MTU*)
3. Unimplemented configuration negotiation, ie, incompatible mismatches
   between client/server configurations are attempted to be detected
   and warned, rather than negotiating settings at connection time.
   (workaround is OCC_{REQUEST/REPLY})

These are essentials that should have been in the code base before initial
release, they are not features that add any value. Instead of implementing
them, the burden of configuration was passed on to the end user, who is
forced to consider these details and attempt a workaround by means of
configuration (all the *mtu, fragment, client-exit variable and ping
timings are made configuration options and left up to the user).