Remove link_mtu environ variable
Due to PMTU discovery, the size of the datagrams sent or receives is
dynamically set during runtime, as network conditions change.
This value was set at --up time, but it makes no sense there, as it is yet
unknown what the network condition to an eventual peer will be.

Removing, rather than setting it to something "harmless" will hopefully
trigger bugs in user scripts which depended on this value before, and they
will be redesigned to not depend on a fixed startup link_mtu value.

This value was also used as command line argument in --up and --down
scripts. The argument was replaced with an empty argument ("") for
backwards compatibility with existing scripts.

Scripts that did not use the value will continue to function correctly,
but scripts that did use it will need to be reworked.