Add MTU frame unit test (mtu.c)
The MTU frame feature is used to calculate buffer sizes for input and
output for both link and tun sides, as a variable of discovered PMTU,
startup configuration (--cipher, --no-replay, --no-iv), and build
configuration (ENABLE_LZO, ENABLE_FRAGMENT)

Only the "new" MTU calculations are checked, as required for integration
of the RFC4821 PMTU discovery algo. This implementation is enabled as
IMPLEMENTED_mtu2 conditional in configure.ac. The legacy calculations are
detected as IMPLEMENTED_mtu_2_0 and IMPLEMENTED_mtu_2_4, but there are no
tests for them, so in code branches without the RFC4821 integration, the
mtu test will be skipped.

The sanity test verifies that the current build configuration calculates
buffer sizes correctly.

In maintainer mode, more combinations of ENABLE_LZO, ENABLE_FRAGMENT, as
well as NDEBUG and ENABLE_DEBUG are tested.