Expect the buffer kichen-sink test to fail
This test fails for me due to an apparent bug in dmalloc triggered by
cmocka (https://github.com/drok/ovpn/issues/7) so it is marked as XFAIL

This will allow the test suite (make check) to continue past the failing
test.

At some point in the future, dmalloc will be fixed, and somone may run
into an unexpected pass. When this happens, please revert this commit,
and to the commit message the version of dmalloc that is fixed, and
cherry-pick it into the tests repo.

It is possible that even after dmalloc is fixed, the test will still fail,
but due to an actual bug in the buffer implementation. To prepare for
this eventuality, a genuine implementation bug will be reported as
a HARD_ERROR, so the XFAIL mark will be ignored, and the test suite will
report the bug. When dmalloc is fixed, and this commit is reverted,
the test will no longer be XFAIL (expected to fail), and any bugs will
be reported as normal errors, not HARD_ERRORs.