Add "does-it-build" integration test
The "does-it-build" test verifies than an otherwise innocuous source
change does not break compilation of unrelated build configurations.

For example, it prevents compilation breakage that requires later fixups.
Examples:
- breakage in --disable-crypto build, fixed later with 51d4d15
- breakage in --disable-server build, fixed later with 8929a39
- breakage in --disable-server build, fixed later with c20434a
- breakage in --disable-crypto build, fixed later with 5acd2ba
- and others.

The 'does-it-build' test builds the codebase simultaneously with various
combinations of ENABLE_* conditionals, making it obvious that a
contemplated patch would break the build in some configurations, and thus
requiring that the breakage be fixed before the the would-be commit is
even committed. (to detect the breakage, the package would have to be
configured for maintainer test mode, --enable-unit-tests=all, and
'make check' would have to be run before attempting to commit)

The test is very basic, it only requires that (on all releases):
- an executable can be produced, and
- the binary does run with the --version command line argument
- running with --version results in a successful exit code (0)

Specifically for the openvpn/2.0 release branch:
- running with --version results in an exit code (1). Normally this flag
  should not cause a failure (non-zero exit), but at least this particular
  release intentionally exits with a failure, so that is deemed to be a
  test "pass"