Port buffer tests to tests/unit framework
This is a complete port, supporting 2.0 to 2.5 implementations of the
buffer feature; However, there are no tests applicable to 2.0 functionality

2.0 and earlier tests: 0
2.1 and earlier tests: 0
2.2 and earlier tests: 8
2.3 and earlier tests: 8 (at commit time, 3/8 are passing, buggy source)
2.4 and earlier tests: 9
2.5 and earlier tests: 11

At 2.3 source the 5 failures appear to be the same bug, ie, implementation
does not collapse/aggregate the list into one buffer. Maybe the bugfix can
be backported from commit 463afd. However, that commit changes the
behaviour of the buffer_list, and it's not clear if this creates a bug
elsewhere where old behaviour is assumed, or if the
buffer_list_aggregate functionality was never used before.