Fix ENABLE_PLUGIN conditionals
This commit properly removes plugin-related code when ENABLE_PLUGIN is not
defined.

Previously, even with ENABLE_PLUGIN undefined, the code would depend on
values defined within $(top_srcdir)/include/openvpn-plugins.h, and
implemented a stub function plugin_defined() that always returned 0

This was a dirty solution, and the tree would not compile without
openvpn_plugins.h being included in the search path.

This commit removes all traces of plugin code from the input to the
compiler when the pre-processor is run without ENABLE_PLUGIN. Access to
openvpn-plugins.h is no longer needed for compile.