Expect all openvpn versions fail --version
The will-it-build integration test, initially written for the 2.0 source
assumed that openvpn --version returning non-zero was a fluke, as it is
unusual. Non zero exit codes by convention indicate some error occurred,
and querying a program's version is not unexpected, nor does it cause any
actual errors.

It appears openvpn does not adhere to this convention in any version,
so the test was changed to understand a return code of "1" as "normal"

This commit changes the expectation that openvpn would return zero/success
on --version invocations.

Derivatives/forks are still expected to return success on --version