Port does-it-build to 2.3 source
The list of source files were split according to the revision they were
introduced at, up to release 2.3

Some conditionals/mutators were renamed from USE_{LZO,CRYPTO,SSL} to
ENABLE_{LZO,CRYPTO,SSL}, they were also ported to be compatible. When the
tests repo is merged into 2.0/2.1/2.2, the USE_* mutators will be used,
and when merged int later releases, their ENABLE_* equivalents are used.

A second SSL library was added in 2.3, PolarSSL, and the build accounts
for this. Existing integration tests will be built with OpenSSL. A separate
commit will add tests for PolarSSL. I'm not adding them now, because
PolarSSL was renamed to mbedtls in the meanwhile, so the 2.3 source needs
to be patched to use mbedtls as an alternative to PolarSSL before the
integration test can be added.

mutator and source lists for 2.1 and 2.2 were added by inspection of git
history, but were not tested. They may need to be adjusted when tests
repo is ported to those release branches