Document build untested build mutators
Some preprocessor defines (source mutators) exist in config.h but are not
exercised in the current integration/does-it-build suite.

These new mutators are also be disabled in the does-it-build tests.

Maybe it would be a good idea to implement a build fuzzer, to activate a
random selection of these otherwise untested mutators, to see what breaks.