Add -ldl explicitly in does-it-build test
Some of the test builds use plugin and OpenSSL, but omitted to include
@DL_LIBS@ on their library list. However, since OpenSSL itself depends on
-ldl, linking was successful anyway, as the dependency was satisfied
by simply including -lssl

However, when linking with PolarSSL, which does not depend on -ldl, the
otherwise identical builds failed with unresolved dl symbols.

This commit explicitly adds -ldl where needed, as future proofing. If ever
the does-it-build test suite is linked against an OpenSSL without engines
it would fail to build. This commit pre-empts that. The right thing is to
explicitly include dependencies, and not rely (accidentally) on them being
implicilty included.