Redesign frame/MTU calculations
mtu.c/mtu.h are redesigned in order to exactly account for the buffer space
needed at each step in the vpn pipeline (reading TUN, compression,
encryption, MTU probing). Each buffer's headroom, overhead and payload are
taken into account.

Previously, the MTU calculations were on a worst case, eg, all buffers were
sized identically, with the generic BUFSIZE(frame).

This one-size-fits-all approach is incompatible with Path MTU discovery
algo, which aims to exactly detect the Path MTU, and also needs to be able
to send frames larger than BUFSIZE(), up to 64k, in order to detect PMTU
that is rising (eg, if a network is 'healing').

Additionally, for MTU management, we'll need to send ICMP messages back to
sender in some cases. Detecting when an MTU is exceeded, and an ICMP needs
to be sent, requires knowing the PMTU exactly. Consider the case of the
multi server, with several clients connected, each with a different PMTU.
In this case, each client has it's own PMTU, although they all share a
TUN buffer and a LINK buffer (ie, the link's frame was used to size all
the buffers). In this case, there is no single frame that accurately
contains info about the individual connection's PMTU.

Additionally, with the RFC4821, receive buffers are decoupled from transmit
buffers, and not derived from one struct frame.