Implement IS_INITIALIZED() macro
This commit implements a simple way to detect use-before initialization
bugs, and allows the check to be removed in production builds, ie where
ASSERT()s are removed by setting -DNDEBUG

This macro relies on a struct member named is_init of type boolean. When
these structs are initialized, the responsible init() function should set
this member to true. Then, related setters and getters can ASSERT()
on the struct being IS_INITIALIZED(mydata) where mydata is the pointer to
the struct given by the caller.