Add scaffolding for backwards-compatible tests
The naive way to maintain a test suite is to commit tests together with
the source files into the same source tree and call it a day. This devalues
the test suite, as it cannot be used to back-test old releases against
bugs found in the future. By maintaining the test suite separate from the
source, it can evolve and be improved with new tests when bugs are found,
and the up-to-date test suite can be merged back to any old source branch
to test it for the new bugs, and to prove that back-ported bugfixes do
work as intended.