Add --enable-unit-tests=all configuration
The default unit-test mode is to tests units relevant for the current
configuration, like a package maintainer would need to do at build
time. This is a _sanity_ verification, rather than a full test, its
goal is to check that the codebase does not interact badly (in an
obvious way) with the libraries on the target system it's being
targetted for.

This commit adds a regression unit test mode for the project
maintainers and patch authors, which will run _all_ available
unit tests on various configurations, exhaustively, whether or not
they are currently active.

This means before maintainers make commits, they can verify easily
how their changes affect the various possible build variants. For
instance, the tls_crypt test would be built and tested with all
supported SSL libraries, with and without ENABLE_DEBUG, with and
without ASSERT's, and with and without ENABLE_CRYPTO. These are all
relevant configuration options for tls_crypt, and are all tested.

Additionally, in maintainer mode, the test suite can/will build
and run test binaries under valgrind, with memory leaks and
uninitialized memory access tracking, to generate meaningful error
logs, and to avoid introducing avoidable bugs into the tree.

Other specialized tools can be required for the maintainer mode
that would be unreasonable to be build dependencies for package
maintainers. In fact some of these verification tools may not be
available on all OS's where the application is supported, but the
maintainers can have a means to do heavy duty testing, even if
end-users are unable to duplicate all of it.