Generate own config-version.h for integration tests
This commit duplicates the package's generation of config-version.h, and
will generate a local one for each integration test.

A duplicate config-version.h enables tests to be run immediately after
configuration, ie, without running "make" in the top builddir first. So,
even if the chosen configuration fails to build for some reason, the tests
can be run to debug the reason for the build failure.