Handle failing static-analysis more graciously
The MTU test performs static analysis of the required headers as a form of
forward bug prevention. This test checks that required pre-processor
defines are set when they should be, and not set when unneccesary, to avoid
future pitfalls where they are used is some part of the codebase unguarded.

On the OpenVPN codebase, OPT_P_PEER_ID is present even when ENABLE_SSL is
not defined, although it makes no sense for it, and other pull-time options
to be set unless SSL is enabled. As a result, over time, the code base has
acquired much dead code and with it, the potential for bugs introduced in
the future. For example, everything in push.c is never called when
ENABLE_SSL is not set, so this is all dead code in this scenario.

This commit marks the tests that fail due to static analysis in OpenVPN as
XFAIL, and reconfigures some tests (NDEBUG and ENABLE_DEBUG) to also
ENABLE_SSL (so as to maximize test coverage)

When the codebase is cleaned-up with respect to ENABLE_SSL (ie, dead code
#ifdef'd away) this commit can be reverted.

If the codebase is cleaned up by removing the ENABLE_SSL option altogether,
the commit can be reverted, but additionally, a backwards compatibility
fixup must be added to test.c to define ENABLE_SSL conditional on
IMPLEMENTED_x_y where x, y represent the revision of the codebase where
ENABLE_SSL was removed (ie, the same test can still run on older codebases
with the full benefit of the static analysis tests)