Add --no-static-analysis to mtu unit test
When the static analysis test fails, the mtu test does normally run the
functional tests.

Also, when run as a sanity test, the purpose of the test is only to verify
that bugs do not exist in the present instance of the code, and is not
concerned with the quality of the codebase, and the potential for future
bugs. To reflect this purpose, when compiled without -DUNIT_TESTS_ALL,
the static analysis tests will not run. They still run in maintainer mode,
where code quality should be a concern.

This commit allows the mtu unit test to skip static analysis and run the
functional tests only, by adding --no-static-analysis to the test's
command line. It also disables static analysis altogether for sanity tests.