Move PAYLOAD_ALIGN define to buffer.c
The only purpose of this define is to set the machine word aligmnent, so
it can be checked if VERIFY_ALIGNMENT is defined. Prior to the commit,
it was defined in mtu.h, but its value is not logically derived from any
mtu parameters. Rather, MTU-related buffer headroom  would need to depend
on buffer alignment requirements. This define logically belongs in buffer.h

This commit moves PAYLOAD_ALIGN to buffer.h