Workaround for automake E2BIG bug
automake 1.13.1d to 1.14.1 contain an E2BIG bug
(http://debbugs.gnu.org/18744) in the detection of gnu make, instroduced at
automake 3de278 (1.13.1d), and fixed e7c14d (1.15)

This definition of am__is_gnu_make overrides the buggy version in that range
of automakes. For versions below the 1.13.1d this define will be unused,
and benign.

This commit adds the fix from the automake repo, to allow integrity tests
Makefiles to build and run with any automake version (tested with 1.11.1)