Fix buffer.c link error with -DDMALLOC
Commit e719a05 introduced a new function, gc_addspecial() but ommited to
provide the _debug version of the function, for use with -DDMALLOC

Found with tests/unit/buffer unit test

The incomplete commit was added before v2.4_alpha1, but is still
incomplete as of v2.4.8. It would have caused any build with -DDMALLOC to
be broken. This is evidence that dmalloc is not in use, and is not truly
a supported debug mode. The developers should consider either using it, or
removing it.

The new tests repo does use DMALLOC to check for memory errors, but in only
in a limited way as of this commit (every unit test is built and run with
dmalloc as one of the configurations, but dmalloc's features are not used
to aggressively hunt for memory problems, only it's default settings are
used)

This commit adds the _debug version to gc_addspecial() and provides the
macros needed to call the substitute _debug function when built with
-DDMALLOC. It fixes commit e719a05