Fix segmentation fault in multi.lzo-compstub.combo
When running  multi.lzo-compstub.combo --version, lzo_version_string() is
called in options.c, but this function is not declared, and is implied
an integer. On 64-bit hosts, instead of a 64-bit pointer to string, the
compiler takes the return value as a 32-bit integer and sign-extends it
to 64-bit pointer to string. Since the address of the version string is
in static memory, it is not mapped like most pointers on the stack at
0xffffff so the resulting address is wrong, causing a segmentation fault.

The bug occurs only with "-DENABLE_LZO -DENABLE_LZO_STUB" but not with
"-DENABLE_LZO". In the first instance there is a compiler warning about
implicit declaration of function 'lzo_version_string', but in the second
there is no warning, as lzo.h includes lzutil.h if LZO_STUB is not
defined.

lzo.h was included into options.c haphazardly via options.h.
Moved #include <lzo.h> from options.h to options.c - it is needed by the
implementation of "options", not by the interface of "options"

Since this bug would have been prevented if warnings were heeded, the next
commit also promotes implicit-function-declarations to errors in the test
suite.

This commit fixes a segmentation fault in --version when compiled with
ENABLE_LZO and ENABLE_LZO_STUB