Remove dead code in next_connection_retry()
... in order to remove unused-but-set-variable error.

This code was introduced at 3796f9 to "fix a bug in proxy fallback", and
partially removed at af1bf85 to "add option --management-query-proxy".

It feels like af1bf85 introduced a new bug by removing the breakout of the
while (!ce_defined) loop. However, there is no test for this code, and the
code is non-trivial, so I can't tell by looking whether a bug is there or
not. In any case, the code being removed was dead, so by removing it I'm
not making the bug worse.