Port sitnl integration test to tests repo
Although it is ported, this test is not yet tested. The porting is done
at codebase 2.4, but this test is not enabled until 2.5

This untested porting is done to keep this test from running (and failing)
in the 2.4 codebase.

It should be tested when the tests repo is integrated into 2.5