Port t_lpback.sh test to tests repo
This test was historically used as a package test (the UUT was the package
build per configure specifications), but it lived in the tests subdir, and
refered to the $top_builddir/src/openvpn binary from within the test. This
caused a dependency issue. The package had to be built before make check
could be run.

The awkward referencing of the built binary from the test dir, and the
dependency issue were clues that logically this test was misfiled.

The t_lpback.sh checks the target binary succeeds when running
with --crypto <x> --test-crypto

The misfiling is that this test is both an integration test, and a package
test. It can be run against the package as configured (if crypto is enabled
in the configuration), and it can be run against relevant integration
targets like those in does-it-build, to check that all possible
integrations are well behaved.

It runs as a package test if --enable-unit-tests=yes or =all, and it runs
as integration test if --enable-unit-tests=all

This commit enables --test-crypto both against the configured and built
package (as a package test), and against selected integration test builds
(p2p, multi combo, multi client, with openssl and with mbedssl).