Implement XFAIL_TESTS configure setting
The "influential variable" XFAIL_TESTS can be set on the configure
command line or on "make check XFAIL_TESTS=..." command line to mark
specific tests as expected to fail.

This will be useful for developers working on a system with known bugs,
such as cmocka or dmalloc bugs, which are not easily fixable. By allowing
specific tests to fail without causing the test suite to fail, they can
continue doing work even in the presence of such a "blind spot".

If this variable is not set at configure or at build time, the individual
Makefiles set the variable to tests known to fail, but due to real bugs,
TODO items, and such.

If XFAIL_TESTS is set on the command line, it overrides the setting of
the individual unit tests. Thus, an easy way to disable XFAILS globally, is
to unset it on the command line (make check XFAIL_TESTS=).

Alternately, a developer working on a buggy system should include *all*
tests expected to fail, ie, even those already included by individual
unit tests on the configure command line.

This commit allows overriding XFAIL_TESTS waivers, and allows development
on a buggy system.