Remove XFAIL waiver from buffer unit test
This kitchen-sink test fails on CentOS6/7 due to an apparent interraction
between cmocka and dmalloc.

This waiver is not neccesary for an updated system, such as Ubuntu bionic.
Thus, since the test is known to pass on Ubuntu, it is enabled.

Also, on travis, this avoids the test suite failing due to XPASS.

The commit enables travis-ci to pass