Disable travis MSVC build
When building on travis, on windows hosts, build-deps.sh clones an
out-of-tree repo at https://github.com/openvpn/openvpn-build.git and
runs the tip of the default branch there, which at the time of this
writing is broken, breaking the build.

Being out of tree, when those scripts break, they break the ovpn build
silently. The breakage will only be noticed the next time code is commited
in ovpn, but will not be related to the commit.

Commenting out the MSVC build until I can integrate the build scripts into
the ovpn tree.