Refactor does-it-build test
In order to allow one set of built binaries to be tested by multiple tests,
the binaries and the tests are decoupled. Binaries are build without
regard to which tests they will be subjected to (ie, their name no longer
contains .combo, .client, etc extension).

From the build binaries, some are selected to be tested as p2p, and/or as
multiserver, and/or as multiclient, depending on their capabilities.

Furthermore, each role will be tested in a subdirectory named after the
role to be tested, such that the logfiles generated for each test will not
conflict. The automake test facility replaces the test extension with .log
so a binary tested with two different roles (eg, as multi.crypto2.client
and multi.crypto2.server, that is to say, the multi.crypto2 tested as
.client and as .server) would result in two log files overwriting each
other. By separating tests by role into subdirectories, the logfiles will
be server/multi.crypto2.log and client/multi.crypto2.log

Removed the ".combo" test role, because I can't think of a useful test.
Such a test would imply verifying that the same binary can satisfy a
--mode combo. Instead, a combo binary can be added to both server_tests
and client_tests to test it in both roles, against a 'reference' peer.
A 'combo' test could check that the binary can interoperate with another
instance of itself, but that test case is covered by including the
'reference' binary to both server_tests and client_tests.

This commit makes the does-it-build test more flexible, allowing the built
binaries to be tested by multiple LOG_COMPILERs